+++ java/swe1/MyServer/src/main/java/at/fhtw/bif3/domain/card/KrakenCard.java
@@ -5,7 +5,7 @@ public class KrakenCard extends Card {
     public KrakenCard() {
     }
 
-    public KrakenCard(String id, String name, int damage, ElementType elementType) {
+    public KrakenCard(String id, String name, double damage, ElementType elementType) {
         super(id, name, damage, elementType);
     }
