+++ java/swe1/MyServer/src/main/java/at/fhtw/bif3/domain/card/WizardCard.java
@@ -5,7 +5,7 @@ public class WizardCard extends Card {
     public WizardCard() {
     }
 
-    public WizardCard(String id, String name, int damage, ElementType elementType) {
+    public WizardCard(String id, String name, double damage, ElementType elementType) {
         super(id, name, damage, elementType);
     }
